requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0
langchain==0.1.9
langchain-openai==0.0.5
langchain-core==0.1.27
PyPDF2==3.0.1
faiss-cpu==1.7.4
pydantic==2.6.1 